make a helper instead of a viewmodel as no view is associated.
